24.07.02 유저 게시판 - 게시물 수정,삭제 프로세스 세션 유저정보 검증 기능 구현 (이진규)

+++ seoulmate/src/main/java/seoulmate/board/UserDeleteController.java
@@ -8,6 +8,8 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import seoulmate.membership.MemberDTO;
+import utils.JSFunction;
 
 @WebServlet("/userdelete.do")
 public class UserDeleteController extends HttpServlet {
@@ -15,6 +17,14 @@ public class UserDeleteController extends HttpServlet {
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
+        
+        MemberDTO member = (MemberDTO) request.getSession().getAttribute("user");
+        if (member == null) {
+            // 유저 정보가 없는 경우 로그인 페이지로 리디렉션
+            JSFunction.alertBack(response, "로그인이 필요한 기능입니다. 해당 게시물은 작성자 및 관리자만 삭제 할 수 있습니다.");
+            return;
+        }
+
         String idx = request.getParameter("idx");
         if (idx == null || !idx.matches("\\d+")) {
             response.sendError(HttpServletResponse.SC_BAD_REQUEST, "Invalid idx parameter");
@@ -22,6 +32,14 @@ public class UserDeleteController extends HttpServlet {
         }
 
         UserBoardDAO dao = new UserBoardDAO();
+        String postAuthor = dao.getPostAuthor(idx);
+
+        // 유저 권한 확인
+        if (member.getUSER_NUM() > 4 && !member.getUSER_ID().equals(postAuthor)) {
+            JSFunction.alertBack(response, "게시글 삭제 권한이 없습니다. 해당 게시물은 작성자 및 관리자만 삭제 할 수 있습니다.");
+            return;
+        }
+
         int result = dao.deletePost(idx);
 
         response.setContentType("text/html; charset=UTF-8");
